+++ common-core/src/main/java/cn/featherfly/common/lang/asserts/IllegalArgumentAssert.java
@@ -2,15 +2,13 @@
 package cn.featherfly.common.lang.asserts;
 
 /**
- * <p>
- * IllegalArgumentAssert
- * 类的说明放这里
- * </p>
+ * IllegalArgumentAssert.
+ *
  * @since 1.7
  * @version 1.7
  * @author zhongj
  */
-public class IllegalArgumentAssert extends LocalizedAssert<IllegalArgumentException>{
+public class IllegalArgumentAssert extends LocalizedAssert<IllegalArgumentException> {
 
     /**
      */